Add integrations tests structure and move json tests to it
